requirements.txt
# run: pip install -i requirements.txt
gitpython==3.1.27
pybedtools==0.9.0
pysam==0.16.0